==> 4_Grid_search.ipynb <==
# Grid search

In this section, we will create a huge pipeline which takes the raw data as an input, concatenates it, preprocesses it, trains a model, predics, and finally scores it. It will also use grid search to tune the parameters. Let's get into it, shall we?

import sys
sys.path.append('scripts')

%matplotlib inline

import random as rnd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sb
import process
import transforms
from multiprocessing import Pool, Queue
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.metrics import f1_score
from sklearn.externals import joblib

sb.set_style('dark')

First let's load the data. We'll load both the training and the testing dataset.

data_train, target_train = process.load_dataset_target('data/train.csv', 'data/train_target.csv')

data_val, target_val = process.load_dataset_target('data/valid.csv', 'data/valid_target.csv')

data = pd.concat([data_train, data_val])
target = pd.concat([target_train, target_val])

target = target[0]

Now we have the data, let's create the pipe.

pipe = make_pipeline(
    transforms.LowVarianceRemover(),
    PCA(0.95),
    transforms.BoxcoxTransform(),
    StandardScaler(),
    ExtraTreesClassifier(n_jobs=-1, verbose=2)
)
pipe.steps

And now the model. We'll use a cross-validated grid search.

%%time
model = GridSearchCV(pipe, dict(pca__n_components=[0.9, 0.95], extratreesclassifier__max_depth=[1, 2, 3]), n_jobs=-1, verbose=2, error_score=0)
model.fit(data, target)

%%time
macro = f1_score(target, model.predict(data), average='macro')
micro = f1_score(target, model.predict(data), average='micro')
print('Best model score: %f' % model.best_score_)
print('Macro F1 score: %f' % macro)
print('Micro F1 score: %f' % micro)

Oh, by the way, what were the best parameters found by grid search?

model.best_params_

Now let's save it.

joblib.dump(model, 'models/extratrees.pkl')
