==> 5_Ensemble_learning.ipynb <==
# Ensemble learning

In the chapter, we will look into ensemble learning. We will train multiple models and hopefully improve out scoring. To keep things quick and simple, we'll be using only minimal preprocessing. This will reduce the overall performance of the models but also decrease the time it takes to train the models.

import sys
sys.path.append('scripts')

%matplotlib inline

import random as rnd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sb
import process
import transforms
from multiprocessing import Pool, Queue
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.preprocessing import StandardScaler
from sklearn.feature_selection import VarianceThreshold
from sklearn.decomposition import PCA
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import BaggingClassifier, VotingClassifier
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.metrics import f1_score
from sklearn.externals import joblib

sb.set_style('dark')

## Data

This should come as no suprise by now. We'll load the training and test dataset.

%%time
data_train, target_train = process.load_dataset_target('data/train.csv', 'data/train_target.csv')
data_val, target_val = process.load_dataset_target('data/valid.csv', 'data/valid_target.csv')

target_train = target_train[0]
target_val = target_val[0]

## Training a model

It's time to make a pipeline and train a proper model. Onward, brothers!

pipeline_bayes = make_pipeline(
    VarianceThreshold(),
    transforms.BoxcoxTransform(),
    StandardScaler(),
    BaggingClassifier(GaussianNB(), n_jobs=-1, verbose=2)
)
pipeline_tree = make_pipeline(
    VarianceThreshold(),
    transforms.BoxcoxTransform(),
    StandardScaler(),
    BaggingClassifier(DecisionTreeClassifier(max_depth=1, class_weight='balanced'), n_jobs=-1, verbose=2)
)

%%time
model = VotingClassifier([
        ('bayes', pipeline_bayes),
        ('tree', pipeline_tree),
    ], n_jobs=-1).fit(data_train, target_train)

## Scoring on the validation dataset

Now let's see how well does the model perform.

%%time
score = model.score(data_val, target_val)

print('Accuracy: %f' % score)

And that's the score of the model. Finally, let's save this model.

joblib.dump(model, 'models/ensemble.pkl')
