==> 6_Conslusion.ipynb <==
# Conclusion

In this awesome data science series, we have looked at my playground where we loaded and analyzed the [Netcla dataset](http://www.neteye-blog.com/netcla-the-ecml-pkdd-network-classification-challenge/). We trained a couple of basic prediction models, then we also trained a few ensemble models.

In order to save our precious time, we always trained only on a sample of the whole dataset. That's probably why the scores were pretty low for most models. We also lacked the domain knowledge to fully understand the data. Therefore, preprocessing was only basic and it further lowered the scores of the models.

The last thing we shall do is to load our saved models, predict labels for the validation data and run the `eval.py` script to see our performance.

import sys
sys.path.append('scripts')

import pandas as pd
import process
from sklearn.externals import joblib

data, _ = process.load_dataset_target('data/valid.csv', 'data/valid_target.csv')

## Notes

These are my personal notes for this task. I must say, it was a lot of fun! I enjoyed writing Python code, transforming the data, and trying various predictors and seeing how they perform. I lacked domain knowledge and therefore didn't bother much with feature engineering and looking into data properly. For instance, I didn't care about outliers.

When I started, I wrote 2 transforms, one of which is the `BoxcoxTransform` (in file `scripts/transforms.py`) which applies boxcox tranform to the dataset. Unfortunately, I wrote it bad, it performed boxcox transforms for each row, instead for each column. These also involved a huge performance penalty.

Of course, I didn't know, so I always used only a small subset of the whole dataset for training models. After I fixed `BoxcoxTransform`, suddenly things moved very quickly. For example, the extra trees model originally trained for 25 minutes using a small subset of the dataset. After the fix, it only trained for 16 seconds!

Amazed by this new found POWER, I trained on the whole dataset using more parameters for grid search and more complex models. Even more fun!

Finally, thank you for this class. It was AMAZING!!!

## Extra trees with grid search

First, predict and store the predictions.

trees_model = joblib.load('models/extratrees.pkl')
trees_result = pd.DataFrame(trees_model.predict(data))

trees_result.to_csv('predictions/extratrees.csv', header=False, index=False)

Now evaluate the model.

%%bash
python3 scripts/eval.py data/valid_target.csv predictions/extratrees.csv

## Ensemble

First, predict and store the predictions.

ensemble_model = joblib.load('models/ensemble.pkl')
ensemble_result = pd.DataFrame(ensemble_model.predict(data))

ensemble_result.to_csv('predictions/ensemble.csv', header=False, index=False)

Now evaluate the model.

%%bash
python3 scripts/eval.py data/valid_target.csv predictions/ensemble.csv

## Evaluation

I don't really now why both models perform so poorly. The extra trees model even does not work, it always predicts 0.

One possible reason for this is that the input data is not being transformed by the model before predicting.

Anyway, I'm signing off. Have a good day!
