==> 7_Balance_dataset.ipynb <==
# Balance the dataset

I forgot that the dataset is so inbalanced. That's the reason I've got so bad models. I could not help myself but try to balance the dataset and train one more model. Let's go.

import sys
sys.path.append('scripts')

%matplotlib inline

import random as rnd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sb
import process
from transforms import LowVarianceRemover, BoxcoxTransform
from multiprocessing import Pool, Queue
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.feature_selection import VarianceThreshold
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import ExtraTreesClassifier, BaggingClassifier, VotingClassifier
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import f1_score
from sklearn.externals import joblib

sb.set_style('dark')

data = process.load_dataset('data/train.csv', 'data/train_target.csv')

non_zero = data[data.target != 0]
non_zero.shape

avg_per_class = non_zero.groupby(by='target').count()[[0]].mean()[0]
avg_per_class

Now we know the following things:

- after removing 0 targets, we are left with about 185000 data points
- there are 9725 data points for each class on average

zero = data[data.target == 0].sample(int(avg_per_class / 3 * 2))

balanced_data = pd.concat([non_zero, zero])
balanced_data.head()

balanced_data.shape

Now we have a balanced dataset. Let's train a model.

## Extra trees

pipeline = make_pipeline(
    LowVarianceRemover(),
    PCA(0.95),
    BoxcoxTransform(),
    StandardScaler(),
    ExtraTreesClassifier(n_jobs=-1, verbose=2)
)

%%time
model = GridSearchCV(pipeline, dict(extratreesclassifier__max_depth=[1, 2, 3]), n_jobs=-1, verbose=2, error_score=0)
model.fit(balanced_data.drop('target', axis=1), balanced_data.target)

model.best_params_

Now let's evaluate the model.

val_data, _ = process.load_dataset_target('data/valid.csv', 'data/valid_target.csv')

predicted = pd.DataFrame(model.predict(val_data))
predicted.to_csv('predictions/extratrees.csv', header=False, index=False)

%%bash
python3 scripts/eval.py data/valid_target.csv predictions/extratrees.csv

## Bagging

pipeline_ens_bagging = make_pipeline(
    VarianceThreshold(),
    BoxcoxTransform(),
    StandardScaler(),
    BaggingClassifier(n_jobs=-1, verbose=2)
)
pipeline_ens_bayes = make_pipeline(
    LowVarianceRemover(),
    PCA(0.95),
    BoxcoxTransform(),
    StandardScaler(),
    GaussianNB()
)

%%time
model_ens = VotingClassifier([
    ('bagging', pipeline_ens_bagging),
    ('bayes', pipeline_ens_bayes)
])
model_ens.fit(balanced_data.drop('target', axis=1), balanced_data.target)

Now let's evaluate this model, too.

predicted_ens = pd.DataFrame(model.predict(val_data))
predicted_ens.to_csv('predictions/ensemble.csv', header=False, index=False)

%%bash
python3 scripts/eval.py data/valid_target.csv predictions/ensemble.csv

To be honest, I expected a better improvement. But I don't know what else to improve, so that's it.
